tiao/requirements/common.txt
python-telegram-bot==11.1.0
expiringdict==1.1.4
requests==2.22.0
Flask==1.1.1
python-decouple==3.1
nltk==3.4.5
pymongo==3.11.0
